+++ src/main/groovy/ru/vyarus/gradle/plugin/animalsniffer/report/ReportMessage.groovy
@@ -13,6 +13,7 @@ class ReportMessage {
     String signature
     String source
     String line
+    String field
     String code
     boolean parseFail
 }