+++ src/test/resources/ru/vyarus/gradle/plugin/animalsniffer/java/field/Sample.java
@@ -1,0 +1,15 @@
+package invalid;
+
+import org.slf4j.LoggerFactory;
+
+import java.nio.file.Path;
+
+public class Sample {
+
+    private Path field;
+
+    public static void main(String[] args) {
+        // method added in 1.7
+        Boolean.compare(true, true);
+    }
+}